+++ fields/core.ts
@@ -12,6 +12,7 @@ const inputs = {
   markdown: normalizeLineBreaks,
   code: null,
   datetime: (v: string) => v ? new Date(v) : null,
+  current_datetime: (v: string) => v ? new Date(v) : null,
   date: null,
   time: null,
   hidden: null,