+++ static/components/f-current_datetime.js
@@ -2,23 +2,19 @@ import { toLocal } from "./utils.js";
 import { Input } from "./f-text.js";
 
 customElements.define(
-  "f-datetime",
+  "f-current_datetime",
   class extends Input {
     get inputAttributes() {
       return { type: "datetime-local", class: "input" };
     }
 
     get value() {
-      const value = super.value;
-
-      if (value) {
-        return format(new Date(value));
-      }
-      return null;
+      return format(new Date());
     }
   },
 );
 
+// Get the value in the format "YYYY-MM-DDTHH:MM"
 function format(date = new Date()) {
   return toLocal(date).toISOString().slice(0, 16);
 }